+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 0.0.6 (2019-04-17)
 
 * enhancement: aesthetics
+* bugfix: buttons should not submit the form
 
 
 Version 0.0.5 (2019-04-17)